% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1479 rows and 7 columns.
}
\usage{
data1
}
\description{
Parent lipids: We have identified parent lipids from serum,
and user can use the results  we provide,
or upload your own identified or downloaded parent lipid results from other databases,
in the template format of "Parent lipids".
}
\examples{
  head(data1)
}
\keyword{datasets}
