% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data2}
\alias{data2}
\title{data2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 94 rows and 27 columns.
}
\usage{
data2
}
\description{
EpiReactions: The information for the 94 epi-metabolic reactions that may occur in lipids included the reaction type,
reaction formula, the change of formula and mass after the reaction,
and the matching relationship between the lipid structure and reactions.
}
\examples{
  head(data2)
}
\keyword{datasets}
