% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data3}
\alias{data3}
\title{data3}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 759 rows and 6 columns.
}
\usage{
data3
}
\description{
allfeatures dataset: Please upload (differential) ion feature data in the template format of "allfeatures dataset"
}
\examples{
  head(data3)
}
\keyword{datasets}
