% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_reaction_enrichment.R
\name{epi_reaction_enrichment}
\alias{epi_reaction_enrichment}
\title{Epi-reaction_enrichment}
\usage{
epi_reaction_enrichment(result, x_value, data2)
}
\arguments{
\item{result}{The result of the lipid_epimetabolite_matching function.
or differential features dataset: Please upload significant changed lipid epimetablites information you detected in a comparison of two conditions,
in the template format of "marker lipid epi-metabolites"}

\item{x_value}{Number of lipid metabolites identified by metabolic reaction (x value)}

\item{data2}{EpiReactions: The information for the 94 epi-metabolic reactions that may occur in lipids included the reaction type,
reaction formula, the change of formula and mass after the reaction,
and the matching relationship between the lipid structure and reactions.}
}
\value{
list of plot and table
}
\description{
Epi-reaction_enrichment to identify and plotting the significantly changed structural modification type of lipids in a disease.
}
\examples{
library(readxl)
library(Lipidepifind)
data_folder <- system.file("extdata", package = "Lipidepifind")
result <- read_xlsx(file.path(data_folder, "result.xlsx")) \%>\% as.data.table()
x_value <- read_xlsx(file.path(data_folder, "x value.xlsx")) \%>\% as.data.table()
data2 <- read_xlsx(file.path(data_folder, "Epireactions.xlsx")) \%>\% as.data.table()
epi_reaction_enrichment_list <- epi_reaction_enrichment(result,x_value,data2)
}
\references{
1. Xinguang Liu, Ang Zhang, Yang Xie, Yanmin Shi, Jianya Yang, Yan Du, Jinyan Wu, Yitong Gao, Jiansheng Li.
Lipidepifind: A Workflow for the Characterization of Lipid Epi-metabolites as Diagnostic Biomarkers for Idiopathic Pulmonary Fibrosis, submitted.
}
