% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_epimetabolite_matching.R
\name{lipid_epimetabolite_matching}
\alias{lipid_epimetabolite_matching}
\title{lipid_epimetabolite_matching}
\usage{
lipid_epimetabolite_matching(
  data1,
  data2,
  data3,
  engine = c("data.table", "sqlite")
)
}
\arguments{
\item{data1}{Parent lipids: We have identified parent lipids from serum,
and user can use the results  we provide,
or upload your own identified or downloaded parent lipid results from other databases,
in the template format of "Parent lipids".}

\item{data2}{EpiReactions: The information for the 94 epi-metabolic reactions that may occur in lipids included the reaction type,
reaction formula, the change of formula and mass after the reaction,
and the matching relationship between the lipid structure and reactions.}

\item{data3}{allfeatures dataset: Please upload (differential) ion feature data in the template format of "allfeatures dataset"}

\item{engine}{"data.table","sqlite"}
}
\value{
result data frame
}
\description{
lipid_epimetabolite_matching is a tool to mz matching of lipid epi-metabolites (structure modifications out of classical bio-pathway).
}
\examples{
library(readxl)
library(Lipidepifind)
data_folder <- system.file("extdata", package = "Lipidepifind")
data1 <- read_xlsx(file.path(data_folder, "Parent lipids.xlsx")) \%>\% as.data.table()
data2 <- read_xlsx(file.path(data_folder, "Epireactions.xlsx")) \%>\% as.data.table()
data3 <- read_xlsx(file.path(data_folder, "Allfeatures dataset.xlsx")) \%>\% as.data.table()
lipid_epimetabolite_matching(data1,data2,data3,engine = "sqlite")

}
\references{
1. Xinguang Liu, Yunfan Zhao, Yang Xie, Yanmin Shi, Jianya Yang, Yan Du, Jinyan Wu, Yitong Gao, Ang Zhang, Jiansheng Li.
Characterization of Lipid Epi-metabolites/reaction as Diagnostic Biomarkers for Idiopathic Pulmonary Fibrosis by Lipidepifind, submitted.
}
