% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_value}
\alias{x_value}
\title{x_value}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63 rows and 2 columns.
}
\usage{
x_value
}
\description{
Number of lipid metabolites identified by metabolic reaction (x value)
}
\examples{
  head(x_value)
}
\keyword{datasets}
